backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
 'passlib==1.7.4',
 'PyJWT==2.8.0',
 'requests==2.31.0',
 'python-dotenv==1.0.0',
 "flask>=3.1.0",
 "flask-cors>=5.0.1",
 "bcrypt>=4.3.0",
]



